+++ src/LaraQuiz/database/migrations/2017_07_04_124200_create_quiz_runner_survey_sessions_table.php
@@ -57,9 +57,9 @@ class CreateQuizRunnerSurveySessionsTable extends Migration
         Schema::create($this->table, function (Blueprint $table) {
             $table->engine = 'InnoDB';
 
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('survey_id');
-            $table->unsignedBigInteger($this->userForeignKey);
+            $table->increments('id');
+            $table->unsignedInteger('survey_id');
+            $table->unsignedInteger($this->userForeignKey);
             $table->timestamp('created_at')->nullable();
             $table->timestamp('ended_at')->nullable();
