+++ src/LaraQuiz/database/migrations/2017_07_04_124400_create_quiz_runner_session_answer_values_table.php
@@ -43,8 +43,8 @@ class CreateQuizRunnerSessionAnswerValuesTable extends Migration
         Schema::create($this->table, function (Blueprint $table) {
             $table->engine = 'InnoDB';
 
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('survey_session_answer_id');
+            $table->increments('id');
+            $table->unsignedInteger('survey_session_answer_id');
             $table->text('value');
 
             $table->unique('survey_session_answer_id');
